Lagt til reponseobjekt for transaksjonskallene

+++ src/main/kotlin/no/nav/bidrag/reskontro/dto/consumer/ReskontroConsumerOutput.kt
@@ -1,6 +1,6 @@
 package no.nav.bidrag.reskontro.dto.consumer
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class ReskontroConsumerOutput(
     val innParametre: ReskontroConsumerInput,
@@ -17,18 +17,18 @@ data class Aktivitet(
     val fodselsOrgNr: String? = null,
     val navn: String? = null,
     val dato: String? = null,
-    val belop: BigInteger? = null
+    val belop: BigDecimal? = null
 )
 
 data class BarnISak(
     val fodselsnummer: String? = null,
-    val restGjeldOffentlig: BigInteger? = null,
-    val restGjeldPrivat: BigInteger? = null,
-    val sumIkkeUtbetalt: BigInteger? = null,
-    val sumForskuddUtbetalt: BigInteger? = null,
-    val restGjeldPrivatAndel: BigInteger? = null,
-    val sumInnbetaltAndel: BigInteger? = null,
-    val sumForskuddUtbetaltAndel: BigInteger? = null,
+    val restGjeldOffentlig: BigDecimal? = null,
+    val restGjeldPrivat: BigDecimal? = null,
+    val sumIkkeUtbetalt: BigDecimal? = null,
+    val sumForskuddUtbetalt: BigDecimal? = null,
+    val restGjeldPrivatAndel: BigDecimal? = null,
+    val sumInnbetaltAndel: BigDecimal? = null,
+    val sumForskuddUtbetaltAndel: BigDecimal? = null,
     val periodeSisteDatoFom: String? = null,
     val periodeSisteDatoTom: String? = null,
     val stoppUtbetaling: String? = null
@@ -36,9 +36,9 @@ data class BarnISak(
 
 data class Bidragssak(
     val bidragssaksnummer: Long,
-    val bmGjeldFastsettelsesgebyr: BigInteger,
-    val bmGjeldRest: BigInteger,
-    val bpGjeldFastsettelsesgebyr: BigInteger,
+    val bmGjeldFastsettelsesgebyr: BigDecimal,
+    val bmGjeldRest: BigDecimal,
+    val bpGjeldFastsettelsesgebyr: BigDecimal,
     val perBarnISak: List<BarnISak>? = emptyList()
 )
 
@@ -48,15 +48,15 @@ data class GjeldendeBetalingsordning(
     val kildeNavn: String? = null,
     val datoSisteGiro: String? = null,
     val datoNesteForfall: String? = null,
-    val belop: BigInteger? = null,
+    val belop: BigDecimal? = null,
     val datoSistEndret: String? = null,
     val aarsakSistEndret: String? = null,
-    val sumUbetalt: BigInteger? = null
+    val sumUbetalt: BigDecimal? = null
 )
 
 data class NyBetalingsordning(
     val datoFraOgMed: String? = null,
-    val belop: BigInteger? = null
+    val belop: BigDecimal? = null
 )
 
 data class Retur(
@@ -66,12 +66,12 @@ data class Retur(
 
 data class Skyldner(
     val fodselsOrgnr: String? = null,
-    val sumLopendeBidrag: BigInteger,
+    val sumLopendeBidrag: BigDecimal,
     val statusInnkrevingssak: String? = null,
     val fakturamaate: String? = null,
     val sisteAktivitet: String? = null,
-    val innbetBelopUfordelt: BigInteger,
-    val gjeldIlagtGebyr: BigInteger
+    val innbetBelopUfordelt: BigDecimal,
+    val gjeldIlagtGebyr: BigDecimal
 )
 
 data class Transaksjon(
@@ -81,9 +81,9 @@ data class Transaksjon(
     val dato: String? = null,
     val kildeFodselsOrgNr: String? = null,
     val mottakerFodslesOrgNr: String? = null,
-    val opprinneligBeloep: BigInteger? = null,
-    val restBeloep: BigInteger? = null,
-    val valutaOpprinneligBeloep: BigInteger? = null,
+    val opprinneligBeloep: BigDecimal? = null,
+    val restBeloep: BigDecimal? = null,
+    val valutaOpprinneligBeloep: BigDecimal? = null,
     val valutakode: String? = null,
     val bidragssaksnummer: Long,
     val periodeSisteDatoFom: String? = null,
